Added showMap(), hideMap() and attached them to related buttons.  Fixed ReminderServiceStub.kt and changed Reminder.kt to take Double for LAT and LONG values.

+++ app/src/main/java/com/example/remindmelater/dto/Reminder.kt
@@ -1,8 +1,6 @@
 package com.example.remindmelater.dto
 
-import retrofit2.http.Body
-
-data class Reminder(var title : String = "", var body: String = "", var latitude : Float = 0.0f, var longitude : Float = 0.0f, var radius : Int = 0, var userEmail : String = "") {
+data class Reminder(var title : String = "", var body: String = "", var latitude : Double = 0.0, var longitude : Double = 0.0, var radius : Int = 0, var userEmail : String = "") {
     override fun toString(): String {
         return "$title $body $userEmail"
     }
